+++ packages/xo-web/src/xo-app/backup-ng/index.js
@@ -5,11 +5,12 @@ import Icon from 'icon'
 import React from 'react'
 import SortedTable from 'sorted-table'
 import StateButton from 'state-button'
-import { confirm } from 'modal'
-import { map, groupBy } from 'lodash'
 import { Card, CardHeader, CardBlock } from 'card'
+import { confirm } from 'modal'
 import { constructQueryString } from 'smart-backup'
 import { Container, Row, Col } from 'grid'
+import { get } from 'xo-defined'
+import { isEmpty, map, groupBy, some } from 'lodash'
 import { NavLink, NavTabs } from 'nav'
 import { routes } from 'utils'
 import {
@@ -30,6 +31,17 @@ import New from './new'
 import FileRestore from './file-restore'
 import Restore from './restore'
 import Health from './health'
+import { destructPattern } from './utils'
+
+const Ul = props => <ul {...props} style={{ listStyleType: 'none' }} />
+const Li = props => (
+  <li
+    {...props}
+    style={{
+      whiteSpace: 'nowrap',
+    }}
+  />
+)
 
 const _runBackupNgJob = ({ id, name, schedule }) =>
   confirm({
@@ -98,6 +110,34 @@ const SchedulePreviewBody = ({ item: job, userData: { schedulesByJob } }) => (
   </table>
 )
 
+const MODES = [
+  {
+    label: 'rollingSnapshot',
+    test: job =>
+      some(job.settings, ({ snapshotRetention }) => snapshotRetention > 0),
+  },
+  {
+    label: 'backup',
+    test: job =>
+      job.mode === 'full' && !isEmpty(get(() => destructPattern(job.remotes))),
+  },
+  {
+    label: 'deltaBackup',
+    test: job =>
+      job.mode === 'delta' && !isEmpty(get(() => destructPattern(job.remotes))),
+  },
+  {
+    label: 'continuousReplication',
+    test: job =>
+      job.mode === 'delta' && !isEmpty(get(() => destructPattern(job.srs))),
+  },
+  {
+    label: 'disasterRecovery',
+    test: job =>
+      job.mode === 'full' && !isEmpty(get(() => destructPattern(job.srs))),
+  },
+]
+
 @addSubscriptions({
   jobs: subscribeBackupNgJobs,
   schedulesByJob: cb =>
@@ -131,11 +171,15 @@ class JobsTable extends React.Component {
         default: true,
       },
       {
-        itemRenderer: _ => (
-          <span style={{ textTransform: 'capitalize' }}>{_.mode}</span>
+        itemRenderer: job => (
+          <Ul>
+            {MODES.filter(({ test }) => test(job)).map(({ label }) => (
+              <Li key={label}>{_(label)}</Li>
+            ))}
+          </Ul>
         ),
         sortCriteria: 'mode',
-        name: _('jobMode'),
+        name: _('jobModes'),
       },
       {
         component: SchedulePreviewBody,
@@ -146,19 +190,19 @@ class JobsTable extends React.Component {
           const { concurrency, offlineSnapshot } = job.settings[''] || {}
 
           return (
-            <ul style={{ listStyleType: 'none' }}>
+            <Ul>
               {concurrency > 0 && (
-                <li>{_.keyValue(_('concurrency'), concurrency)}</li>
+                <Li>{_.keyValue(_('concurrency'), concurrency)}</Li>
               )}
               {offlineSnapshot && (
-                <li>
+                <Li>
                   {_.keyValue(
                     _('offlineSnapshot'),
                     <span className='text-success'>{_('stateEnabled')}</span>
                   )}
-                </li>
+                </Li>
               )}
-            </ul>
+            </Ul>
           )
         },
         name: _('formNotes'),